refactor: extract aura serialization and threat engine runner to shared locations
- Create @wow-threat/shared/src/aura-serialization.ts with
  serializeInitialAurasByActor and deserializeInitialAurasByActor, removing
  identical implementations from events.ts, client-threat-engine.ts, and
  threat-engine.worker.ts
- Create apps/web/src/lib/threat-engine-runner.ts with runThreatEngineForFight,
  consolidating the duplicated fight-lookup → config-resolve → engine-process
  pipeline that existed in both the main-thread fallback and the dedicated worker
