+++ apps/web/src/workers/threat-engine.worker.ts
@@ -1,10 +1,10 @@
 /**
  * Dedicated worker for threat engine event processing.
  */
-import { resolveConfigOrNull } from '@wow-threat/config'
-import { ThreatEngine, buildThreatEngineInput } from '@wow-threat/engine'
+import { ThreatEngine } from '@wow-threat/engine'
 import type { WCLEvent } from '@wow-threat/wcl-types'
 
+import { runThreatEngineForFight } from '../lib/threat-engine-runner'
 import {
   loadThreatWorkerRawEventChunks,
   saveThreatWorkerProcessedResult,
@@ -19,50 +19,6 @@ import type {
 } from './threat-engine-worker-types'
 
 const threatEngine = new ThreatEngine()
-
-function deserializeInitialAurasByActor(
-  initialAurasByActor: Record<string, number[]> | undefined,
-): Map<number, readonly number[]> {
-  if (!initialAurasByActor) {
-    return new Map()
-  }
-
-  return Object.entries(initialAurasByActor).reduce(
-    (result, [actorId, auraIds]) => {
-      const parsedActorId = Number.parseInt(actorId, 10)
-      if (!Number.isFinite(parsedActorId)) {
-        return result
-      }
-
-      const sanitizedAuraIds = auraIds
-        .filter((auraId) => Number.isFinite(auraId))
-        .map((auraId) => Math.trunc(auraId))
-      if (sanitizedAuraIds.length === 0) {
-        return result
-      }
-
-      result.set(
-        parsedActorId,
-        [...new Set(sanitizedAuraIds)].sort((left, right) => left - right),
-      )
-      return result
-    },
-    new Map<number, readonly number[]>(),
-  )
-}
-
-function serializeInitialAurasByActor(
-  initialAurasByActor: Map<number, readonly number[]>,
-): Record<string, number[]> {
-  return Object.fromEntries(
-    [...initialAurasByActor.entries()]
-      .filter(([, auraIds]) => auraIds.length > 0)
-      .map(([actorId, auraIds]) => [
-        String(actorId),
-        [...new Set(auraIds)].sort((left, right) => left - right),
-      ]),
-  )
-}
 
 function toErrorMessage(error: unknown): string {
   if (error instanceof Error) {
@@ -130,59 +86,21 @@ function processThreatEvents(params: {
   rawEventCount: number
 } {
   const { payload, rawEvents, startedAt } = params
-  const fight = payload.report.fights.find(
-    (candidateFight) => candidateFight.id === payload.fightId,
-  )
-  if (!fight) {
-    throw new Error(`fight ${payload.fightId} not found in report payload`)
-  }
-
-  const config = resolveConfigOrNull({
+  const processedPayload = runThreatEngineForFight({
+    engine: threatEngine,
+    fightId: payload.fightId,
+    inferThreatReduction: payload.inferThreatReduction,
+    initialAurasByActor: payload.initialAurasByActor,
+    rawEvents,
     report: payload.report,
+    startedAt,
+    tankActorIds: payload.tankActorIds,
   })
-  if (!config) {
-    throw new Error(
-      `no threat config for gameVersion ${payload.report.masterData.gameVersion}`,
-    )
-  }
 
-  const initialAurasByActor = deserializeInitialAurasByActor(
-    payload.initialAurasByActor,
-  )
-  const { actorMap, friendlyActorIds, enemies, abilitySchoolMap } =
-    buildThreatEngineInput({
-      fight,
-      actors: payload.report.masterData.actors,
-      abilities: payload.report.masterData.abilities,
-    })
-
-  const { augmentedEvents, initialAurasByActor: effectiveInitialAurasByActor } =
-    threatEngine.processEvents({
-      rawEvents,
-      initialAurasByActor,
-      actorMap,
-      friendlyActorIds,
-      abilitySchoolMap,
-      enemies,
-      encounterId: fight.encounterID ?? null,
-      report: payload.report,
-      fight,
-      inferThreatReduction: payload.inferThreatReduction,
-      tankActorIds: new Set(payload.tankActorIds),
-      config,
-    })
-
-  const processDurationMs = Math.round(performance.now() - startedAt)
   return {
-    augmentedEventCount: augmentedEvents.length,
-    processDurationMs,
-    processedPayload: {
-      augmentedEvents,
-      initialAurasByActor: serializeInitialAurasByActor(
-        effectiveInitialAurasByActor,
-      ),
-      processDurationMs,
-    },
+    augmentedEventCount: processedPayload.augmentedEvents.length,
+    processDurationMs: processedPayload.processDurationMs,
+    processedPayload,
     rawEventCount: rawEvents.length,
   }
 }
